requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
oauthlib==3.1.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
PySide6==6.2.2.1
requests==2.27.1
requests-oauthlib==1.3.0
shiboken6==6.2.2.1
toml==0.10.2
tweepy==4.4.0
urllib3==1.26.8
